Ajuste de validación de ip estudiante

+++ src/app/estudiante-list/estudiante-list.component.ts
@@ -4,6 +4,8 @@ import { DatePipe } from '@angular/common';
 import { RouterModule } from '@angular/router';
 import { Estudiante } from '../model/estudiante.interface';
 import { CommonModule } from '@angular/common';
+import { NotaService } from '../services/nota.service';
+import { firstValueFrom } from 'rxjs';
 
 /**
  * Componente para listar estudiantes.
@@ -19,11 +21,12 @@ import { CommonModule } from '@angular/common';
 export default class EstudianteListComponent {
   // Inyección de servicio para manejar operaciones con estudiantes
   private estudianteService = inject(EstudianteService);
-  
+  private notaServ = inject(NotaService);
   // Lista de estudiantes
   lstestudiantes: Estudiante[] = [];
   successMessage: boolean = false;
   successText: string='';
+  successDelete: boolean=false;
 
   /**
    * Método de ciclo de vida que se ejecuta al inicializar el componente.
@@ -48,19 +51,37 @@ export default class EstudianteListComponent {
    * Muestra una confirmación antes de proceder con la eliminación.
    * @param estudiante - Estudiante a eliminar.
    */
-  deleteEstudiante(estudiante: Estudiante) {
+  async deleteEstudiante(estudiante: Estudiante) {
     const confirmacion = window.confirm('¿Está seguro que desea eliminar el registro?'); // Solicita confirmación
     if (confirmacion) {
-      // Si se confirma, se llama al servicio para eliminar el estudiante
-      this.estudianteService.delete(estudiante.id).subscribe(() => {
+      const notas = await firstValueFrom(this.notaServ.list());
+
+      for (let nota of notas) {
+        if (nota.estudiante === estudiante.id) {
+          this.successDelete = true;
+          break;
+        }
+      }
+      if (this.successDelete) {
+        this.successDelete = false;
         this.successMessage = true;
+        this.successText = 'Posee notas asociadas al estudiante seleccionado. Debe borrar o reasignar los registros para continuar.';  // Asignamos el mensaje de actualización
+        setTimeout(() => {
+          this.successMessage = false; // Ocultar el mensaje después de 3 segundos
+          this.loadAll(); // Recarga la lista de estudiantes
+        }, 3000);
+        console.log('No se puede eliminar'); // Mensaje de confirmación en la consola
+      } else {
+        this.estudianteService.delete(estudiante.id).subscribe(() => {
+          this.successMessage = true;
           this.successText = 'El registro ha sido eliminado exitosamente.';  // Asignamos el mensaje de actualización
           setTimeout(() => {
             this.successMessage = false; // Ocultar el mensaje después de 3 segundos
             this.loadAll(); // Recarga la lista de estudiantes
           }, 3000);
-        console.log('ok'); // Mensaje de éxito
-      });
+          console.log('ok'); // Mensaje de confirmación en la consola
+        });
+      }
     } else {
       console.log('Eliminación cancelada'); // Mensaje de cancelación
     }
